pyproject.toml
[tool.poetry]
name = "works"
version = "0.1.0"
description = ""
authors = ["Islam Ayman <islam.ayman.mohammed@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
